services/consumer/requirements.txt
kafka-python
clickhouse-driver
prometheus-client>=0.19.0
python-json-logger>=2.0.7
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-kafka-python>=0.41b0
opentelemetry-exporter-jaeger>=1.20.0
avro>=1.11.3
jsonschema>=4.20.0
minio
